bao ani oi

+++ core/src/com/finalproject/game/client/packet/client/MouseMove.java
@@ -2,13 +2,12 @@ package com.finalproject.game.client.packet.client;
 
 public class MouseMove {
 
-    public int mouseX;
-    public int mouseY;
-
+    public float mouseX;
+    public float mouseY;
 
     public MouseMove() {}
 
-    public MouseMove(int mouseX, int mouseY) {
+    public MouseMove(float mouseX, float mouseY) {
         this.mouseX = mouseX;
         this.mouseY = mouseY;
     }
